Integrate serve_cli with ray scripts
